split the debugger package into main and debugger
