+++ packages/extension-adapter/src/index.ts
@@ -1,24 +1,21 @@
-import { Accessor, createEffect, createSignal, onCleanup } from "solid-js"
+import { createEffect, createSignal, onCleanup } from "solid-js"
+import { registerDebuggerPlugin, PluginFactory } from "@solid-devtools/debugger"
 import {
 	MESSAGE,
 	onWindowMessage,
 	postWindowMessage,
 	startListeningWindowMessages,
 } from "@shared/messanger"
-import type { SerialisedTreeRoot, BatchUpdateListener } from "@shared/graph"
+import type { SerialisedTreeRoot } from "@shared/graph"
 import { getArrayDiffById } from "./handleDiffArray"
 
 startListeningWindowMessages()
 
-export function useExtensionAdapter({
+const extensionAdapterFactory: PluginFactory = ({
 	forceTriggerUpdate,
-	roots,
+	serialisedRoots,
 	makeBatchUpdateListener,
-}: {
-	forceTriggerUpdate: VoidFunction
-	makeBatchUpdateListener: (listener: BatchUpdateListener) => VoidFunction
-	roots: Accessor<SerialisedTreeRoot[]>
-}): { enabled: Accessor<boolean> } {
+}) => {
 	const [enabled, setEnabled] = createSignal(false)
 
 	postWindowMessage(MESSAGE.SolidOnPage)
@@ -29,7 +26,7 @@ export function useExtensionAdapter({
 
 	// diff the roots array, and send only the changed roots (edited, deleted, added)
 	createEffect((prev: SerialisedTreeRoot[]) => {
-		const _roots = roots()
+		const _roots = serialisedRoots()
 		const diff = getArrayDiffById(prev, _roots)
 		postWindowMessage(MESSAGE.GraphUpdate, diff)
 		return _roots
@@ -38,4 +35,18 @@ export function useExtensionAdapter({
 	makeBatchUpdateListener(updates => postWindowMessage(MESSAGE.BatchedUpdate, updates))
 
 	return { enabled }
+}
+
+/**
+ * Registers the extension adapter with the debugger.
+ */
+export function useExtensionAdapter() {
+	registerDebuggerPlugin(data => {
+		const { enabled } = extensionAdapterFactory(data)
+		return {
+			enabled,
+			trackSignals: enabled,
+			trackBatchedUpdates: enabled,
+		}
+	})
 }
