+++ packages/main/package.json
@@ -1,7 +1,7 @@
 {
-	"name": "@solid-devtools/locator",
-	"version": "0.4.2",
-	"description": "On-page overlay for localisating SolidJS components, and revealing them in your code editor.",
+	"name": "solid-devtools",
+	"version": "0.4.3",
+	"description": "Runtime helpers for hooking up SolidJS application with Solid Devtools",
 	"license": "MIT",
 	"author": "Damian Tarnawski <gthetarnav@gmail.com>",
 	"contributors": [],
@@ -16,16 +16,17 @@
 	"keywords": [
 		"solid",
 		"devtools",
-		"components",
-		"locator",
-		"vscode"
+		"reactivity"
 	],
 	"private": false,
 	"sideEffects": true,
-	"type": "module",
+	"publishConfig": {
+		"access": "public"
+	},
 	"files": [
 		"dist"
 	],
+	"type": "module",
 	"main": "./dist/server.cjs",
 	"module": "./dist/server.js",
 	"types": "./dist/index.d.ts",
@@ -44,25 +45,18 @@
 	"scripts": {
 		"dev": "tsup --watch",
 		"build": "tsup",
-		"test": "echo \"NOOP TEST\"",
+		"test": "echo NOOP TEST",
 		"typecheck": "tsc --noEmit"
 	},
 	"devDependencies": {
-		"esbuild": "^0.14.43",
-		"esbuild-plugin-solid": "^0.4.2",
 		"solid-js": "^1.4.4",
 		"tsup": "^6.1.2",
 		"typescript": "^4.7.3"
 	},
 	"dependencies": {
 		"@solid-devtools/debugger": "^0.0.1",
-		"@solid-devtools/ui": "0.2.0",
-		"@solid-primitives/bounds": "^0.0.100",
-		"@solid-primitives/event-listener": "^2.2.0",
-		"@solid-primitives/keyboard": "^0.0.100",
-		"@solid-primitives/platform": "^0.0.100",
-		"@solid-primitives/utils": "^2.1.1",
-		"motion": "^10.10.0"
+		"@solid-devtools/locator": "^0.4.2",
+		"@solid-devtools/extension-adapter": "^0.4.0"
 	},
 	"peerDependencies": {
 		"solid-js": "^1.4.4"
