+++ packages/main/src/index.ts
@@ -1,0 +1,7 @@
+import { useExtensionAdapter } from "@solid-devtools/extension-adapter"
+
+// Extension adapter
+useExtensionAdapter()
+
+export * from "@solid-devtools/debugger"
+export * from "@solid-devtools/locator"
