+++ packages/main/src/server.ts
@@ -1,0 +1,2 @@
+export * from "@solid-devtools/debugger"
+export * from "@solid-devtools/locator"
