+++ packages/main/tsup.config.ts
@@ -1,0 +1,3 @@
+import defineConfig from "../../configs/tsup.config"
+
+export default defineConfig("ts", undefined, true)
