+++ playgrounds/sandbox/vite.config.ts
@@ -11,8 +11,9 @@ export default defineConfig({
 		alias: {
 			// used inside packages/debugger
 			"@shared": path.resolve(pathToPackages, "shared"),
+			"solid-devtools": path.resolve(pathToPackages, "main", "src"),
+			"@solid-devtools/debugger": path.resolve(pathToPackages, "debugger", "src"),
 			"@solid-devtools/ui": path.resolve(pathToPackages, "ui", "src"),
-			"solid-devtools": path.resolve(pathToPackages, "debugger", "src"),
 			"@solid-devtools/locator": path.resolve(pathToPackages, "locator", "src"),
 			"@solid-devtools/extension-adapter": path.resolve(pathToPackages, "extension-adapter", "src"),
 		},
