+++ tsconfig.json
@@ -17,7 +17,8 @@
 		"baseUrl": ".",
 		"paths": {
 			"@shared/*": ["./packages/shared/*"],
-			"solid-devtools": ["./packages/debugger/src/index.ts"],
+			"solid-devtools": ["./packages/main/src/index.ts"],
+			"@solid-devtools/debugger": ["./packages/debugger/src/index.ts"],
 			"@solid-devtools/ui": ["./packages/ui/src/index.tsx"],
 			"@solid-devtools/locator": ["./packages/locator/src/index.tsx"],
 			"@solid-devtools/extension-adapter": ["./packages/extension-adapter/src/index.ts"]
